edit parseInt to parseFloat

+++ node-require/calculate.js
@@ -2,8 +2,8 @@ const add = require('./add');
 const subtract = require('./subtract');
 const multiply = require('./multiply');
 const divide = require('./divide');
-const a = parseInt(process.argv[2]);
-const b = parseInt(process.argv[4]);
+const a = parseFloat(process.argv[2]);
+const b = parseFloat(process.argv[4]);
 if (process.argv[3] === 'plus') {
   console.log(`result: ${add(a, b)}`);
 }
